requirements.txt
requests>=2.28.0
pandas>=1.5.0
vaderSentiment>=3.3.2
matplotlib>=3.6.0
seaborn>=0.12.0
python-dotenv>=1.0.0
outscraper>=6.0.0
huggingface-hub>=0.24.0
flask>=3.0.0
gunicorn>=21.0.0
rq>=1.16.0
redis>=5.0.0
sqlalchemy>=2.0.0
